requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
python-multipart==0.0.9
jinja2==3.1.4
pydantic==2.8.2

azure-storage-blob==12.20.0
